useState 분리 완료
